LayoutTests/js/constant-fold-not-nan-expected.txt

Tests that our parse-time constant folding treats !(0/0) correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !(0/0) is true
PASS successfullyParsed is true

TEST COMPLETE
